Update for use with sfdk target
The opal-merge-translations.sh script makes use of lconvert, which for
Sailfish OS development is more likely to be installed in an sfdk target
than in the user's path. It therefore makes sense to execute the script
in the scratbox2 target via sfdk.

Unfortunately the sfdk version of sed doesn't support the -E flag. The
-r flag, which is equivalent, is supported. This change therefore
changes the script to use -r rather than -E.

The README has also been updated to explain how to run the script using
sfdk.